API - Admin Management Mobil

+++ database/migrations/02_users.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->bigInteger('no_sim');
             $table->string('email')->unique();
             $table->string('password');
-            $table->unsignedInteger('role_id');
+            $table->unsignedBigInteger('role_id');
             $table->timestamps();
 
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
